+++ Example/Tests/ReferenceImages/BetterSegmentedControlTests/a_BetterSegmentedControl__its_indicatorViewBackgroundColor_property__when_it_is_set_during_initialization__renders_correctly_after_updating_it@3x.png
