+++ Example/Tests/ReferenceImages/BetterSegmentedControlTests/a_BetterSegmentedControl__its_indicatorViewInset_property__when_it_is_set_during_initialization__renders_correctly@3x.png
