+++ Example/Tests/ReferenceImages/BetterSegmentedControlTests/a_BetterSegmentedControl__its_indicator_view_border_related_properties__when_they_are_set_during_initialization__renders_correctly_after_updating_border_width_to_0@3x.png
