+++ Example/Tests/ReferenceImages/BetterSegmentedControlTests/a_BetterSegmentedControl__when_setting_cornerRadius_and_indicatorViewInset__first_cornerRadius_then_indicatorViewInset__renders_correctly@3x.png
