+++ Example/Tests/ReferenceImages/BetterSegmentedControlTests/a_BetterSegmentedControl__when_setting_cornerRadius_and_indicatorViewInset__first_indicatorViewInset_then_cornerRadius__renders_correctly@3x.png
