+++ Example/Tests/ReferenceImages/BetterSegmentedControlTests/a_BetterSegmentedControl__with_most_basic_initialization__with_three_segments__renders_correctly@3x.png
